Merge Vita controls into port 0
SceDs3 calls sceCtrlRegisterVirtualControllerDriver in its module_start
which sets 0x81009AC0 in SceCtrl to a pointer, pointing to a function
table for interfacing with the virtual controller driver. In the
function at 0x8100107C in SceCtrl this pointer is checked and if the
port is 0, the input control data is discarded except for power button
down.

	ctrl->buttons = *(uint32_t*)0x81009A84
		| *(uint32_t*)0x81009C08
		| ctrl->buttons & 0x40000000;
	ctrl->lx = *(uint8_t*)0x81009A88;
	ctrl->ly = *(uint8_t*)0x81009A94;
	ctrl->rx = *(uint8_t*)0x81009AA0;
	ctrl->ry = *(uint8_t*)0x81009AAC;

We merge the Vita controls by merging the input control data to port 0
with the control data for controller 1 found at these addresses.